fixed the build
lets not repack the whole obfuscated vanilla into our jar, instead we
now repack a jar that only consists of the resources from the default
resource pack.
also this broke the build because for some reason the vanilla main got
invoked.
also updated to java 1.8, because for years people have had it and it
has some interesting features like inline λs.